+++ lib/QoreElementsOperatorNode.cpp
@@ -51,12 +51,12 @@ QoreValue QoreElementsOperatorNode::evalImpl(bool& needs_deref, ExceptionSink* x
         return QoreValue();
 
     switch (v->getType()) {
-        case NT_LIST: return v->get<const QoreListNode>()->size();
+        case NT_LIST: return (int64)v->get<const QoreListNode>()->size();
         // return the number of characters in a string (not bytes)
-        case NT_STRING: return v->get<const QoreStringNode>()->length();
-        case NT_HASH: return v->get<const QoreHashNode>()->size();
+        case NT_STRING: return (int64)v->get<const QoreStringNode>()->length();
+        case NT_HASH: return (int64)v->get<const QoreHashNode>()->size();
         case NT_OBJECT: return v->get<const QoreObject>()->size(xsink);
-        case NT_BINARY: return v->get<const BinaryNode>()->size();
+        case NT_BINARY: return (int64)v->get<const BinaryNode>()->size();
     }
 
     return 0;
